pyproject.toml
[tool.poetry]
name = "adtarasov-mlops"
version = "0.1.0"
description = "MLOps project in MLHS course"
authors = ["alexeydtarasov <alexey.d.tarasov@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.4"
numpy = "^1.26.3"
matplotlib = "^3.8.2"
dvc-gdrive = "^3.0.1"
scikit-learn = "^1.3.2"
pyyaml = "^6.0.1"
fire = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
dvc = "^3.38.1"

[tool.black]
line-length = 90
target-version = ["py37"]
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 90
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
